requirements/base.txt
pytz==2020.5  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.0.1  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.1.0  # https://github.com/celery/django-celery-beat
django-timezone-field==4.0  # https://github.com/celery/django-celery-beat/pull/378
flower==0.9.7  # https://github.com/mher/flower
uvicorn[standard]==0.13.2  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.0.11  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.10.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
django-cors-headers==3.6.0 # https://github.com/adamchainz/django-cors-headers

#Graphene
graphene>=2.0 #https://github.com/graphql-python/graphene
graphene-django==2.13.0
django-graphql-jwt==0.3.1
graphene-django-optimizer==0.7.0
graphql-core==2.3.2
graphql-relay==2.0.1
django-graphiql-debug-toolbar==0.1.4